jbig2dec: Use correct define for maxium value of type.
Fixes Coverity CID 355177.

+++ jbig2dec/jbig2_page.c
@@ -80,11 +80,11 @@ jbig2_page_info(Jbig2Ctx *ctx, Jbig2Segment *segment, const uint8_t *segment_dat
             if (index >= ctx->max_page_index) {
                 /* grow the list */
 
-                if (ctx->max_page_index == SIZE_MAX) {
+                if (ctx->max_page_index == UINT32_MAX) {
                     return jbig2_error(ctx, JBIG2_SEVERITY_FATAL, segment->number, "too many pages in jbig2 image");
                 }
-                else if (ctx->max_page_index > (SIZE_MAX >> 2)) {
-                    ctx->max_page_index = SIZE_MAX;
+                else if (ctx->max_page_index > (UINT32_MAX >> 2)) {
+                    ctx->max_page_index = UINT32_MAX;
                 }
 
                 pages = jbig2_renew(ctx, ctx->pages, Jbig2Page, (ctx->max_page_index <<= 2));
